SUBMARINE-1208. Move files into folders
