+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/ClusterRestApi.java
@@ -23,7 +23,7 @@ import org.apache.submarine.commons.cluster.ClusterServer;
 import org.apache.submarine.commons.cluster.meta.ClusterMeta;
 import org.apache.submarine.commons.cluster.meta.ClusterMetaType;
 import org.apache.submarine.commons.utils.SubmarineConfiguration;
-import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.apache.submarine.server.rest.workbench.annotation.SubmarineApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
