+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/workbench/LoginRestApi.java
@@ -25,7 +25,7 @@ import org.apache.submarine.server.rest.workbench.annotation.SubmarineApi;
 import org.apache.submarine.server.workbench.database.entity.SysUserEntity;
 import org.apache.submarine.server.workbench.database.mappers.SysUserMapper;
 import org.apache.submarine.server.database.utils.MyBatisUtil;
-import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
