+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/workbench/ProjectRestApi.java
@@ -20,8 +20,8 @@ package org.apache.submarine.server.rest.workbench;
 
 import com.github.pagehelper.PageInfo;
 
-import org.apache.submarine.server.utils.JsonResponse;
-import org.apache.submarine.server.utils.JsonResponse.ListResult;
+import org.apache.submarine.server.utils.response.JsonResponse;
+import org.apache.submarine.server.utils.response.JsonResponse.ListResult;
 import org.apache.submarine.server.rest.workbench.annotation.SubmarineApi;
 import org.apache.submarine.server.workbench.database.entity.ProjectEntity;
 import org.apache.submarine.server.workbench.database.service.ProjectService;
