+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/utils/gson/EnvironmentIdDeserializer.java
@@ -17,19 +17,21 @@
  * under the License.
  */
 
-package org.apache.submarine.server.utils;
+package org.apache.submarine.server.utils.gson;
 
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonPrimitive;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-import org.apache.submarine.server.api.notebook.NotebookId;
+import com.google.gson.JsonParseException;
+import org.apache.submarine.server.api.environment.EnvironmentId;
 
 import java.lang.reflect.Type;
 
-public class NotebookIdSerializer implements JsonSerializer<NotebookId> {
+public class EnvironmentIdDeserializer implements JsonDeserializer<EnvironmentId> {
+
   @Override
-  public JsonElement serialize(NotebookId src, Type typeOfSrc, JsonSerializationContext context) {
-    return new JsonPrimitive(src.toString());
+  public EnvironmentId deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+      throws JsonParseException {
+    return EnvironmentId.fromString(json.getAsJsonPrimitive().getAsString());
   }
 }
