+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/utils/gson/ExperimentIdSerializer.java
@@ -16,17 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.submarine.server.utils;
 
-import com.google.gson.ExclusionStrategy;
-import com.google.gson.FieldAttributes;
+package org.apache.submarine.server.utils.gson;
 
-public class JsonExclusionStrategy implements ExclusionStrategy {
-  public boolean shouldSkipClass(Class<?> arg0) {
-    return false;
-  }
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
+import org.apache.submarine.server.api.experiment.ExperimentId;
 
-  public boolean shouldSkipField(FieldAttributes f) {
-    return false;
+import java.lang.reflect.Type;
+
+public class ExperimentIdSerializer implements JsonSerializer<ExperimentId> {
+  @Override
+  public JsonElement serialize(ExperimentId src, Type typeOfSrc, JsonSerializationContext context) {
+    return new JsonPrimitive(src.toString());
   }
 }
