+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/utils/response/JsonExclusionStrategy.java
@@ -16,22 +16,17 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+package org.apache.submarine.server.utils.response;
 
-package org.apache.submarine.server.utils;
+import com.google.gson.ExclusionStrategy;
+import com.google.gson.FieldAttributes;
 
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParseException;
-import org.apache.submarine.server.api.environment.EnvironmentId;
+public class JsonExclusionStrategy implements ExclusionStrategy {
+  public boolean shouldSkipClass(Class<?> arg0) {
+    return false;
+  }
 
-import java.lang.reflect.Type;
-
-public class EnvironmentIdDeserializer implements JsonDeserializer<EnvironmentId> {
-
-  @Override
-  public EnvironmentId deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-      throws JsonParseException {
-    return EnvironmentId.fromString(json.getAsJsonPrimitive().getAsString());
+  public boolean shouldSkipField(FieldAttributes f) {
+    return false;
   }
 }
