+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/utils/response/JsonResponse.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.submarine.server.utils;
+package org.apache.submarine.server.utils.response;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.gson.Gson;
@@ -26,6 +26,12 @@ import com.google.gson.TypeAdapter;
 import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.api.experiment.ExperimentId;
 import org.apache.submarine.server.api.notebook.NotebookId;
+import org.apache.submarine.server.utils.gson.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.gson.NotebookIdDeserializer;
+import org.apache.submarine.server.utils.gson.NotebookIdSerializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
