+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/AbstractSubmarineServerTest.java
@@ -55,10 +55,10 @@ import org.apache.submarine.commons.utils.SubmarineConfVars;
 import org.apache.submarine.server.api.environment.Environment;
 import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.rest.RestConstants;
-import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
-import org.apache.submarine.server.utils.EnvironmentIdSerializer;
-import org.apache.submarine.server.utils.JsonResponse;
 import org.apache.submarine.server.utils.TestUtils;
+import org.apache.submarine.server.utils.gson.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeMatcher;
 import org.junit.Assert;
