+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/EnvironmentRestApiTest.java
@@ -27,9 +27,9 @@ import org.apache.submarine.server.api.environment.Environment;
 import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.api.spec.EnvironmentSpec;
 import org.apache.submarine.server.api.spec.KernelSpec;
-import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
-import org.apache.submarine.server.utils.EnvironmentIdSerializer;
-import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.gson.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
