+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/RegisteredModelRestApiTest.java
@@ -35,8 +35,8 @@ import javax.ws.rs.core.Response;
 import org.apache.submarine.server.api.experiment.ExperimentId;
 import org.apache.submarine.server.model.database.entities.RegisteredModelEntity;
 import org.apache.submarine.server.model.database.service.RegisteredModelService;
-import org.apache.submarine.server.utils.ExperimentIdDeserializer;
-import org.apache.submarine.server.utils.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdSerializer;
 
 public class RegisteredModelRestApiTest {
   private final RegisteredModelService registeredModelService = new RegisteredModelService();
