+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/workbench/SysDeptRestApiTest.java
@@ -21,8 +21,8 @@ package org.apache.submarine.server.rest.workbench;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-import org.apache.submarine.server.utils.JsonResponse;
-import org.apache.submarine.server.utils.JsonResponse.ListResult;
+import org.apache.submarine.server.utils.response.JsonResponse;
+import org.apache.submarine.server.utils.response.JsonResponse.ListResult;
 import org.apache.submarine.server.workbench.database.entity.SysDeptEntity;
 import org.apache.submarine.server.workbench.database.entity.SysDeptTree;
 import org.junit.After;
