+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/workbench/database/utils/DictAnnotationTest.java
@@ -23,8 +23,8 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.reflect.TypeToken;
 
-import org.apache.submarine.server.utils.DictAnnotation;
-import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.response.DictAnnotation;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.apache.submarine.server.rest.workbench.CommonDataTest;
 import org.apache.submarine.server.rest.workbench.SysUserRestApi;
 import org.junit.Test;
