+++ submarine-test/test-k8s/src/test/java/org/apache/submarine/rest/ExperimentRestApiIT.java
@@ -39,11 +39,11 @@ import org.apache.submarine.server.api.experiment.Experiment;
 import org.apache.submarine.server.api.experiment.ExperimentId;
 import org.apache.submarine.server.api.environment.Environment;
 import org.apache.submarine.server.rest.RestConstants;
-import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
-import org.apache.submarine.server.utils.EnvironmentIdSerializer;
-import org.apache.submarine.server.utils.ExperimentIdDeserializer;
-import org.apache.submarine.server.utils.ExperimentIdSerializer;
-import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.gson.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.gson.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.joda.time.DateTime;
 import org.junit.Assert;
 import org.junit.Before;
