+++ submarine-test/test-k8s/src/test/java/org/apache/submarine/rest/NotebookRestApiIT.java
@@ -40,11 +40,11 @@ import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.api.notebook.Notebook;
 import org.apache.submarine.server.api.notebook.NotebookId;
 import org.apache.submarine.server.rest.RestConstants;
-import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
-import org.apache.submarine.server.utils.EnvironmentIdSerializer;
-import org.apache.submarine.server.utils.JsonResponse;
-import org.apache.submarine.server.utils.NotebookIdDeserializer;
-import org.apache.submarine.server.utils.NotebookIdSerializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.gson.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.gson.NotebookIdDeserializer;
+import org.apache.submarine.server.utils.gson.NotebookIdSerializer;
+import org.apache.submarine.server.utils.response.JsonResponse;
 import org.joda.time.DateTime;
 import org.junit.Assert;
 import org.junit.BeforeClass;
